+++ sklearn/cluster/__init__.py
@@ -7,9 +7,9 @@ from .spectral import spectral_clustering, SpectralClustering
 from .mean_shift_ import (mean_shift, MeanShift,
                           estimate_bandwidth, get_bin_seeds)
 from ._affinity_propagation import affinity_propagation, AffinityPropagation
-from .hierarchical import (ward_tree, AgglomerativeClustering, linkage_tree,
-                           FeatureAgglomeration)
-from .k_means_ import k_means, KMeans, MiniBatchKMeans
+from ._hierarchical import (ward_tree, AgglomerativeClustering, linkage_tree,
+                            FeatureAgglomeration)
+from ._k_means import k_means, KMeans, MiniBatchKMeans
 from ._dbscan import dbscan, DBSCAN
 from .optics_ import (OPTICS, cluster_optics_dbscan, compute_optics_graph,
                       cluster_optics_xi)
